Update secret keys location
As 'signing.secretKeyRingFile' property in gradle.properties is
configured to relative secret key ('secring.gpg'), Gradle is unable
to sign artifacts as secret key is stored only in the repo root.
To prevent this, the secret key will be copied to each module which
has to be signed.

Also, as pubring.kbx.asc is not used during signing procedure, it
will be removed.

Affects #52