Remove sunshine.properties
The Attribute interface and its implementations were designed to
retrieve values from sunshine.properties file. As the file is deleted,
all these classes (Attribute and implementations) are deleted.

``tests-regex`` is a system property (former ``tests``) for configuring
a regex for tests filtering.

The name of a temporary directory is hardcoded as it's internal details
of LoadableTestNGSuite.java implementation.

#139