Use tatools.org domain for the project
Now each module, that needs to be released, has to have its own
`gradle.properies` file with two variables:
- NAME provides a short name for the module
- DESCRIPTION gives some explanation of the module

`bintray.gradle` is a generic project's extension which allows
publishing of the artifacts to the bintray repository.

The unit tests for the core module were updated. Most significant change
is that `SunTest` doesn't interrupt whole tests execution for the
module. Previously, as we don't wrap `System.exit` call, it stops tests
execution after the `SunTest` is executed.

#145