Add ability to specify classes as a suite
Now there are two suites (SuiteFromClasses and SuiteFromFileSystem)
which can represent a SunshineSuite. SuiteFromClasses allows creation
a test suite based on predefined classes. Any of these classes can be
used while a custom suite (specific for a test runner) is build.

Also, TestFromClass is created as we need to represent a class as an
instance of SunshineTest.

All relevant documentation and unit tests are updated.

#137